rename trigger-job to trigger-build
